Increase magic effect limit: change from uint8_t to uint16_t
Allow more simultaneous magic effects (uint8_t → uint16_t)

+++ engine/src/game.cpp
@@ -6468,14 +6468,14 @@ void Game::addCreatureHealth(const SpectatorHashSet& spectators, const Creature*
 	}
 }
 
-void Game::addMagicEffect(const Position& pos, uint8_t effect)
+void Game::addMagicEffect(const Position& pos, uint16_t effect)
 {
 	SpectatorHashSet spectators;
 	map.getSpectators(spectators, pos, true, true);
 	addMagicEffect(spectators, pos, effect);
 }
 
-void Game::addMagicEffect(const SpectatorHashSet& spectators, const Position& pos, uint8_t effect)
+void Game::addMagicEffect(const SpectatorHashSet& spectators, const Position& pos, uint16_t effect)
 {
 	for (Creature* spectator : spectators) {
 		if (Player* tmpPlayer = spectator->getPlayer()) {
@@ -8503,8 +8503,8 @@ void Game::updateSpectatorsPvp(Thing* thing)
 	}
 }
 
-bool Game::hasLootType(uint8_t lootTypeId) {
-	for (uint8_t i = LOOT_ARMOR; i <= LOOT_LAST; i++) {
+bool Game::hasLootType(uint16_t lootTypeId) {
+	for (uint16_t i = LOOT_ARMOR; i <= LOOT_LAST; i++) {
 		LootType_t loottype = static_cast<LootType_t>(i);
 		if (loottype == lootTypeId) {
 			return true;
